# src/main/resources/openapi-templates/openapi.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "This is a job scraper application.",
        "title": "Job Scraper API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api/v1",
    "paths": {
        "/jobs": {
            "get": {
                "description": "Get a list of jobs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobScraper"
                ],
                "summary": "Get jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/scraper.JobPosting"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scrape": {
            "post": {
                "description": "Start scraping jobs based on the provided configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobScraper"
                ],
                "summary": "Start scraping",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JobPosting Title",
                        "name": "jobTitle",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Country",
                        "name": "country",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Number of Pages",
                        "name": "pages",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "indeed",
                            "linkedin"
                        ],
                        "type": "string",
                        "description": "Source of job listings (indeed or linkedin)",
                        "name": "source",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "scraper.CompanyDetails": {
            "description": "Company details",
            "type": "object",
            "properties": {
                "industry": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "platform_company_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "scraper.JobPosting": {
            "description": "Job posting details",
            "type": "object",
            "properties": {
                "company_details": {
                    "$ref": "#/definitions/scraper.CompanyDetails"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "platform_job_id": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/scraper.ScraperType"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "scraper.ScraperType": {
            "description": "Type of job scraper",
            "type": "string",
            "enum": [
                "indeed",
                "linkedin"
            ],
            "x-enum-varnames": [
                "Indeed",
                "LinkedIn"
            ]
        }
    }
}